feat: add get REST types for voice state
